refactor: move DevServer-related logic to src/dev/
Move dev server code to its own directory to better reflect the
visle/dev module boundary. createDevManifest and its helpers move to
src/dev/manifest.ts, while createDevLoader and getServerEnvironment
move to src/dev/index.ts. The corresponding tests are co-located.
