+++ test/utils.ts
@@ -6,7 +6,7 @@ import { createBuilder, mergeConfig, type UserConfig } from 'vite'
 
 import { visle } from '../src/build/index.ts'
 import { asRel } from '../src/core/path.ts'
-import { createDevLoader } from '../src/server/dev.ts'
+import { createDevLoader } from '../src/dev/index.ts'
 import { createRender } from '../src/server/render.ts'
 
 const tmpDir = path.resolve('test/__generated__/integration')
